static cast long long to std::size_t. to prevent conversion warning.
/mnt/d/dev/OpenVIII_CPP_WIP/cmake-build-debug-wsl-clang15/_deps/ut_fetch-src/include/boost/ut.hpp:745:27: error: implicit conversion changes signedness: 'long long' to 'std::size_t' (aka 'unsigned long') [-Werror,-Wsign-conversion]
        std::size_t val = std::stoll(argument, &last);
                    ~~~   ^~~~~~~~~~~~~~~~~~~~~~~~~~~
1 error generated.

+++ include/boost/ut.hpp
@@ -742,7 +742,7 @@ struct cfg {
         // parse size argument
         std::size_t last;
         std::string argument(argv[i]);
-        std::size_t val = std::stoll(argument, &last);
+        std::size_t val = static_cast<std::size_t>(std::stoll(argument, &last));
         if (last != argument.length()) {
           std::cerr << "cannot parse option of " << argv[i - 1] << " "
                     << argv[i] << std::endl;
